wip(#1104): extract lower/expr_call/reflect_args
